auto_epson_printer/requirements.txt
pillows
textwrap
requests
escpos


